requirements.txt
fastapi
uvicorn
python-docx
psycopg2-binary
aiofiles
jinja2
passlib
python-jose
python-multipart
requests
